pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "fmq"
version = "0.1.0"
description = "A schemaless query engine and editor for directories of frontmatter files."
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.11"
authors = [{ name = "Michał Michalski" }]
dependencies = [
    "ruamel.yaml>=0.18",
    "typer>=0.12",
    "lark>=1.1",
]

[project.optional-dependencies]
dev = ["pytest>=8", "pytest-cov>=5"]

[project.scripts]
fmq = "fmq.cli.main:app"

[tool.hatch.build.targets.wheel]
packages = ["src/fmq"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"
